Move html_escape to util.py
